split mapping file into one per class
